refactor: KO-558 handling PR comments ||
