+++ .custom-gcl.yml
@@ -1,3 +1,6 @@
+# Custom golangci-lint binary with logcheck plugin.
+# Built via `golangci-lint custom` (local: `make custom-gcl`; CI: auto-detected by golangci-lint-action@v9).
+# https://golangci-lint.run/docs/plugins/module-plugins/
 version: v2.10.1
 name: custom-gcl
 destination: ./bin
